resource authorization handler -> only creator can update and delete restaurant

+++ Services/RestaurantService.cs
@@ -1,8 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantWebApi.Authotization;
 using RestaurantWebApi.Dto;
 using RestaurantWebApi.Entities;
 using RestaurantWebApi.Exceptions;
+using System.Security.Claims;
 
 namespace RestaurantWebApi.Services
 {
@@ -11,16 +14,19 @@ namespace RestaurantWebApi.Services
         private readonly RestaurantContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationSerivce;
 
         public RestaurantService(
             RestaurantContext dbContext,
             IMapper mapper,
-            ILogger<RestaurantService> logger
+            ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService
         )
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationSerivce = authorizationService;
         }
 
         public RestaurantDto GetById(int id)
@@ -50,16 +56,17 @@ namespace RestaurantWebApi.Services
             return restaurantDto;
         }
 
-        public int Create(CreateRestaurantDto createRestaurantDto)
+        public int Create(CreateRestaurantDto createRestaurantDto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(createRestaurantDto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
             return restaurant.Id;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogWarning($"DELETE action on Restaurant IDENTITY: {id} invoked");
 
@@ -71,16 +78,37 @@ namespace RestaurantWebApi.Services
                 );
                 throw new NotFoundException("Restaurant not found");
             }
+            var authorizationResult = _authorizationSerivce
+                .AuthorizeAsync(
+                    user,
+                    restaurant,
+                    new ResourceAutorizationRequirement(ResourceOperation.Delete)
+                )
+                .Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException("Restaurant created by different user");
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
 
-        public void Update(int id, UpdateRestaurantDto dto)
+        public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
         {
             var restaurant = _dbContext.Restaurants.FirstOrDefault(x => x.Id == id);
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationSerivce
+                .AuthorizeAsync(
+                    user,
+                    restaurant,
+                    new ResourceAutorizationRequirement(ResourceOperation.Update)
+                )
+                .Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException("Restaurant created by different user");
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
             restaurant.HasDelivery = dto.HasDelivery;
